Gene Prediction by Exonerate: Perl pipeline.
Gene Prediction by Exonerate: Perl pipeline for prediction of genes
based on Guy Slater’s Exonerate.
As used in Brown et al. Characterisation of retroviruses in the horse
genome and their transcriptional activity via transcriptome sequencing.
Virology in press 2012
and
Tarlinton et al. Characterisation of a group of endogenous
gammaretroviruses in the canine genome. The Veterinary Journal in press
2012